Move all simplification code into one source file
